+++ CHANGELOG.md
@@ -1,6 +1,32 @@
 # Changelog
 
 
+## 0.6.0 (2016-05-14)
+
+* **NEW** 命令行程序支持指定拼音风格:
+
+  ```shell
+  $ pinyin -s Normal 你好
+  ni hao
+  ```
+* **Bugfixed** 解决韵母 i, u, ü 的问题：根据以下拼音方案，还原出正确的韵母
+   [#8](https://github.com/mozillazg/go-pinyin/pull/8),  [python-pinyin#26](https://github.com/mozillazg/python-pinyin/pull/26)
+
+    > i 行的韵母，前面没有声母的时候，写成：yi（衣），yɑ（呀），ye（耶），
+    > yɑo（腰），you（忧），yɑn（烟），yin（因），yɑnɡ（央），yinɡ（英），yonɡ（雍）。
+    >
+    > u 行的韵母，前面没有声母的时候，写成wu（乌），wɑ（蛙），wo（窝），
+    > wɑi（歪），wei（威），wɑn（弯），wen（温），wɑnɡ（汪），wenɡ（翁）。
+    >
+    > ü行的韵母跟声母j，q，x拼的时候，写成ju（居），qu（区），xu（虚），
+    > ü上两点也省略；但是跟声母l，n拼的时候，仍然写成lü（吕），nü（女）。
+
+    **注意** `y` 既不是声母也不是韵母。详见 [汉语拼音方案](http://www.edu.cn/20011114/3009777.shtml)
+
+* **Bugfixed** 解决未正确处理鼻音 ḿ, ń, ň, ǹ 的问题：包含鼻音的拼音不应该有声母
+
+
+
 ## 0.5.0 (2016-03-12)
 
 * **CHANGE** 改为使用来自 [pinyin-data](https://github.com/mozillazg/pinyin-data) 的拼音数据。